+++ NEWS.md
@@ -1,3 +1,5 @@
+- Forcing language identifiers to be lower-case for consistency with ISO conventions.
+
 # psychTestR 2.21.0
 
 - Added `allow_download` argument to audio_NAFC_page.